[feed] Improve GitHub activity feed #26
Small enhancements are added to the activity page such as GitHub octicons
and the corresponding color for each event. Additionally, emojis utilizing
the colon format are replaced with their respective icons. GitHub username
mentions and website links are handled by the ShowdownJS library. A section
is added to the homepage with a preview of the recent development activity.

Closes #26